% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_adjust.R
\name{bh_adjust}
\alias{bh_adjust}
\title{Adjust for Multiple Comparisons}
\usage{
bh_adjust(p)
}
\arguments{
\item{p}{A vector of p-values.}
}
\value{
A vector indicating statistical significance after
adjustment.
}
\description{
Evaluate a vector of p-values for statistical significance on
5\% False Discovery Rate according to the Benjamini-Hochberg
procedure.
}
