% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplots}
\alias{boxplots}
\title{Prepare Boxplots}
\usage{
boxplots(d0, df, fit, help, scl, rt_vars, which = "brains")
}
\arguments{
\item{d0}{Raw data as generated by \code{import_data}}

\item{df}{Pre-processed data as generated by
\code{preprocess_data}}

\item{fit}{A list of linear regressions generated by
\code{fit_regressions}}

\item{help}{A helpers list generated by \code{extract_helpers}}

\item{scl}{Scaling data.frame generated by \code{preprocess_data}}

\item{rt_vars}{Response times variables names generated by
\code{extract_rt_variables}}

\item{which}{A character indicating whether subcortical structures
("brains", default), neuropsychology measures stratified by OSA
("cognition_1") or neuropsychology measures stratified by PD
("cognition_2") should be plotted.}
}
\value{
A list with regressions, one per outcome
}
\description{
Takes in raw and pre-processed data, regression models and
a list of helpers to prepare a selected boxplot.
}
