% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{data_paths}
\alias{data_paths}
\title{Document Paths to Data}
\usage{
data_paths()
}
\value{
A data.frame containing paths to data
}
\description{
This function takes no parameter but needs to be
adjusted manually if desired. It is used in the
\pkg{targets} pipeline to check for changes in data.
}
