% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describes the Sample}
\usage{
describe_data(data)
}
\arguments{
\item{data}{A data.frame generated by \code{import_data}.}
}
\value{
A data.frame containing per group descriptive
statistics.
}
\description{
Takes in data and computes descriptive statistics with
associated statistical tests for variables deemed
descriptive for the paper.
}
