% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extracts Coefficients of Bayesian Regressions}
\usage{
extract_coefficients(fit)
}
\arguments{
\item{fit}{The regression model computed by \code{\link[brms]{brm}}.}
}
\value{
A table summarising models coefficient of interest.
}
\description{
From a Bayesian regression fit, extracts parameter
estimates with associated metrics.
}
