% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{extract_helpers}
\alias{extract_helpers}
\title{Prepare 'Helper' Files}
\usage{
extract_helpers()
}
\value{
A list with data frames containing different
kinds of helper files.
}
\description{
This function takes no parameter but needs to be
adjusted manually if desired.
It is used in the \pkg{targets} pipeline to check
for changes in files that do not contain data proper
but 'helper' data, i.e., data mapping raw data
structures to functions used throughout the pipeline.
}
