% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{extract_rt_variables}
\alias{extract_rt_variables}
\title{Extracts Response Times Variables}
\usage{
extract_rt_variables(help)
}
\arguments{
\item{help}{A list containing data.frame named psych with
a proper study-specific structure.}
}
\value{
A vector of variable names.
}
\description{
Takes in helpers files, finds the neuropsychology helper
and identifies tests in domains that are based on response
times.
The function is very dataset specific!
}
