% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funs.R
\name{fit_bayesian}
\alias{fit_bayesian}
\title{Fit Bayesian Regressions}
\usage{
fit_bayesian(df, formulas)
}
\arguments{
\item{df}{The data.}

\item{formulas}{A two-layered list with formulas for
Bayesian regression.}
}
\value{
A list with regressions, one per outcome.
}
\description{
Loops through outcomes to calculate a set of
Bayesian linear regressions.
}
