% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funs.R
\name{fit_reg}
\alias{fit_reg}
\title{Fit Linear Regressions}
\usage{
fit_reg(d, outcomes, X = "SUBJ * AHI.F + AGE + GENDER + SBTIV", w = FALSE)
}
\arguments{
\item{d}{The data.}

\item{outcomes}{A character vector of the outcomes
of interest.}

\item{X}{The right side of the linear model.}

\item{w}{Indicator whether regression weight shall
be used (TRUE) or not(FALSE, default)}
}
\value{
A list with linear models, one per outcome
}
\description{
Loops through outcomes to calculate a set of
linear regressions.
}
\seealso{
\code{\link[=fit_regressions]{fit_regressions()}} wraps this function.
}
