% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funs.R
\name{fit_regressions}
\alias{fit_regressions}
\title{Fit a Set of Linear Regressions}
\usage{
fit_regressions(df, help)
}
\arguments{
\item{df}{The data frame}

\item{help}{A list with helpers files prepared
by \code{extract_helpers}}
}
\value{
A list with linear models, one per outcome
}
\description{
Loops through outcomes to calculate a set of
linear regressions.
}
\seealso{
\code{\link[=fit_reg]{fit_reg()}} is used to fit single regression lists.
}
