% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_funs.R
\name{freqprop}
\alias{freqprop}
\title{Print Frequency (Proportion)}
\usage{
freqprop(x, d = 0)
}
\arguments{
\item{x}{A vector from which frequencies and proportions
ought to be calculated}

\item{d}{Decimals (defaults to 0)}
}
\value{
A character with frequencies and proportions
}
\description{
Takes in a character/factor vector, calculates
frequency and proportion of its groups and returns
these formatted.
}
