% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{import_data}
\alias{import_data}
\title{Imports Raw Data}
\usage{
import_data(files, helpers)
}
\arguments{
\item{files}{A data.frame containing paths do data
files generated by \code{data_paths}.}

\item{helpers}{A list containing helper files generated
by \code{extract_helpers}}
}
\value{
A tibble containing raw data.
}
\description{
Takes in file paths to data and helper files,
imports the data and structure them properly.
}
\seealso{
\code{\link[=data_paths]{data_paths()}}, \code{\link[=extract_helpers]{extract_helpers()}}
}
