% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_funs.R
\name{lm_coeff}
\alias{lm_coeff}
\title{Extracts Linear Model Coefficients}
\usage{
lm_coeff(fit, term = "SUBJ1:AHI.F1", type = "frequentist")
}
\arguments{
\item{fit}{The regression model computed by \code{lm}
or \code{\link[brms]{brm}}}

\item{term}{Regression coefficient of interest}

\item{type}{A character indicating whether it is
a "frequentist" (default) or "Bayesian" model being
evaluated}
}
\value{
A table summarising models coefficient of interest
}
\description{
From a linear regression fit, extracts parameter estimates
with their associated metrics.
}
