% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_funs.R
\name{lm_dia}
\alias{lm_dia}
\title{Residual Model Diagnostics}
\usage{
lm_dia(fit)
}
\arguments{
\item{fit}{The linear regression model computed by \code{lm}}
}
\value{
A table summarising model diagnostics
}
\description{
From a linear regression fit, extracts measures
of the Breusch-Pagan test for homoscedasticity,
number of outliers according to the Cook distance,
and the Shapiro-Wilk test for normality.
}
