% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{mass}
\alias{mass}
\title{Extracts Effects}
\usage{
mass(fit, type = "moderation")
}
\arguments{
\item{fit}{A list of regression models computed
by \code{lm} or \code{\link[brms]{brm}}}

\item{type}{A character indicating whether only
a "moderation" (default) effects or all comparisons
("full") should be estimated}
}
\value{
A table summarising models coefficient of interest
}
\description{
Using the \code{\link[emmeans]{emmeans}} function
extracts different types of effect estimates from
fitted linear models.
}
