% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{model_check}
\alias{model_check}
\title{Plots Posterior Predictive Checks}
\usage{
model_check(fit, help, formulas, which = "prior_sense")
}
\arguments{
\item{fit}{A nested list with Bayesian regression models}

\item{help}{Helper list generated by \code{extract_helpers}}

\item{formulas}{A nested list with formulas for
Bayesian regression}

\item{which}{A character indicated whether prior
sensitivity tables ("prior_sense", default) or
posterior predictive check ("ppc") ought to be
returned.}
}
\value{
Either a nested list with prior sensitivity
metrics or a list of plots with posterior predictive
checks.
}
\description{
Loops through Bayesian regression models and returns
either prior sensitivity to power scaling table or
posterior predictive check figure.
}
