% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_funs.R
\name{msd}
\alias{msd}
\title{Prints Mean and Standard Deviation}
\usage{
msd(x, d = 2)
}
\arguments{
\item{x}{A vector from which mean and SD ought to be calculated}

\item{d}{Decimals (defaults to 2)}
}
\description{
Takes in a numeric vector, calculates mean and
standard deviation (SD) and returns these formatted.
}
