% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process the Data}
\usage{
preprocess_data(data, help, rt_vars, return = "df")
}
\arguments{
\item{data}{A data.frame containing raw data generated
by \code{import_data}.}

\item{help}{A list containing helper files generated
by \code{extract_helpers}}

\item{rt_vars}{A character of response times variables
generated by \code{extract_rt_variables}}

\item{return}{A character indicating what shall be
returned. Either "df" to return the data or "scl"
to return scaling values.}
}
\value{
A tibble containing raw data.
}
\description{
Takes in raw data, helper files, response times
variables and indicator of the object that shall
be returned.
}
