% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_funs.R
\name{re_formulate}
\alias{re_formulate}
\title{Reformat Model Formula}
\usage{
re_formulate(form)
}
\arguments{
\item{form}{A character containing model formula}
}
\value{
Formated formula for presentation
}
\description{
Takes in a model formula and changes the terms
to presentation-ready format.
}
