% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_funs.R
\name{rprint}
\alias{rprint}
\title{Print Numeric}
\usage{
rprint(x, d = 2)
}
\arguments{
\item{x}{A numeric to be printed}

\item{d}{Decimals (defaults to 2)}
}
\value{
A character with the number
}
\description{
Takes in a number and prints it with specified decimal places.
}
