% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_funs.R
\name{set_formulas}
\alias{set_formulas}
\title{Prepare Formulas}
\usage{
set_formulas()
}
\value{
A list with formulas
}
\description{
Lists formulas for Bayesian models in the
sensitivity-to-heteroscedasticity analysis.
}
