% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statextract.R
\name{statextract}
\alias{statextract}
\title{Extracts and formats test statistics}
\usage{
statextract(coeffs, y, stat = "t")
}
\arguments{
\item{coeffs}{Regression coefficients}

\item{y}{Outcome of interest}

\item{stat}{A character indicating whether t-values
("t", default) or z-values ("z") should be looked for.}
}
\value{
A tibble containing regression coefficients
together with test statistics and p-values associated
with them.
}
\description{
Takes in results of a regression analysis and extracts
regression coefficients with accompanying test statistics
and p-values.
}
