% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_regressions.R
\name{summarise_regressions}
\alias{summarise_regressions}
\title{Summarise a Set of Regressions}
\usage{
summarise_regressions(fit)
}
\arguments{
\item{fit}{A list of linear regression models}
}
\value{
A list with summary of linear models, one per model
}
\description{
Loops through linear regression fits and
summarises their parameters.
}
