% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_funs.R
\name{zerolead}
\alias{zerolead}
\title{Get Rid of Leading Zero}
\usage{
zerolead(x, d = 3)
}
\arguments{
\item{x}{A numeric to be stripped of zero}

\item{d}{Decimals (defaults to 3)}
}
\value{
A character with the number
}
\description{
Takes in a number and prints it with specified
decimal places without the leading zero.
}
