% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Compute Z-scores}
\usage{
zscore(calc, x, nam, AGE, GEN, EDU)
}
\arguments{
\item{calc}{A data.frame with calculator values}

\item{x}{A vector with test scores to be evaluated}

\item{nam}{A data.frame containing mapping of variable names}

\item{AGE}{A vector of observed age values in years}

\item{GEN}{A vector of observed gender values (female = 0, male = 1)}

\item{EDU}{A vector with observed education values in years}
}
\value{
A vector of z-scores
}
\description{
Takes in linear calculator values, observed score,
name of the scale, age, gender and education and returns
a vector of z-scores.
}
